app/requirements.txt
streamlit==1.42.0
azure-ai-documentintelligence==1.0.0
azure-search-documents==11.5.1
openai==1.63.2
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn==0.30.6
protobuf>=3.20.3
pyyaml>=6.0.0
tenacity>=9.0.0
pandas>=2.0.0
transformers>=4.44.2
torch>=2.4.1
datasets>=3.0.1
nltk>=3.8.1
groq